backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.0,<5.0